Refactor to support multi threding
