+++ src/main/java/io/datasearch/diseasedata/store/query/Queries.txt
