+++ src/main/resources/weather-precipitation-data.csv
@@ -4311,7 +4311,11 @@
 01AM0012,Amparai Tank,7.28,81.6,20131023,0
 01AM0012,Amparai Tank,7.28,81.6,20131024,0
 01AM0012,Amparai Tank,7.28,81.6,20131025,0
-01AM0012,Amparai Tank,7.28,81.6,20131026,0
+01AM0012,Amparai Tank,7.28,81.6,20131026,Example : Precipitation data
+
+Raw data
+
+0
 01AM0012,Amparai Tank,7.28,81.6,20131027,0
 01AM0012,Amparai Tank,7.28,81.6,20131028,0
 01AM0012,Amparai Tank,7.28,81.6,20131029,0
